+++ src/pretalx/api/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import include, url
 from rest_framework import routers
 from rest_framework.authtoken.views import obtain_auth_token
 
-from pretalx.api.views import event, review, room, speaker, submission
+from pretalx.api.views import event, review, room, speaker, submission, user
 
 default_router = routers.DefaultRouter()
 default_router.register(r'events', event.EventViewSet)
@@ -18,6 +18,7 @@ event_router.register(r'rooms', room.RoomViewSet)
 app_name = 'api'
 urlpatterns = [
     url(r'^', include(default_router.urls)),
+    url(r'^me$', user.MeView.as_view(), name='user.me'),
     url(r'^auth/', obtain_auth_token),
     url(r'^events/(?P<event>[^/]+)/', include(event_router.urls)),
 ]